Fixing type bug

+++ sportsdirect/boxscore.py
@@ -26,6 +26,6 @@ class BoxScoreFeed(BaseFeed):
     def parse(self, xml_text):
         root = etree.fromstring(xml_text)
         c = root.xpath('//team-sport-content/league-content/season-content/competition')[0]
-        self.home_handicap = int(
+        self.home_handicap = float(
             c.xpath('./betting/point-spread[@closing="true"]/home-handicap/text()')[0])
         return self
